chore: move util methods around + rename utils/server
